Switch Link to v0.1 (was v1).
It is not yet clear that the Link schema is finalized. Open questions:
- Should "_type" be optional?
- Should "products" be removed in favor of "subject"?

Since we are not yet decided, the 0.1 version makes this more clear.

NOTE: This goes against semantic versioning, but no one has depended on
this yet, and we do want to use the 0.x series to indicate "in
development", so it seems like a decent enough tradeoff to make at this
point.

Signed-off-by: SAY-5 <say.apm35@gmail.com>