Support attestation subject URI as alternative for digest
Hello, I am a member of the Google team, working with Mark Lodato, Tom Hennen and others.

Currently, an in-toto Attestation Statement can have multiple subjects, where each subject has the form {name, digest} and where both fields must be present. It is currently impossible to specify a subject that has no meaningful digest.

We would like to support subjects that are not described by a content digest. Examples:

*    A subject that identifies a builder (example: https://host/builder@version). This could be the subject of an attestation that the builder meets a certain SLSA level, and it could be referenced in the builder section of a provenance attestation.
*    A subject that identifies a specific revision of a source-code repository (for example, svn+ssh://host/path/revision-number). This subject could be referenced in the materials section of a provenance attestation, and it could match the subject of an attestation that the repository meets a certain SLSA level.

More background is in #92.