Explain that `goModuleH1` can be used outside go
Previously the spec was written specifically for go modules, which
always have a single top-level directory entry of the form
`name@version`. However, this function is not specific to go and does
not need to have a restriction that there is a single `name@version`
top-level directory.

Now the spec explains that it is designed but not specifically for go
modules, and walks through a detailed example of how to compute. It also
better explains how to run the "equivalent" command.

Signed-off-by: SAY-5 <say.apm35@gmail.com>