Update SPDX predicate specification
This commit contains a suggestion for updating the SPDX predicate
specification, to include the version in the predicateType.

The motivation for this change is same to the motivation that was
mentioned for including the version in the CycloneDX predicate
specification which is that it allows a consumer to determine the
version without having to parse the predicate itself.

Signed-off-by: SAY-5 <say.apm35@gmail.com>