fix(docs): we generate multiple types of code
The docs claimed that "it's typical to keep generated Go code in the
repository itself", only --- we generate multiple types of code (Go,
Java, and Python at present). Make the statement more ambiguous to
convey that we will store multiple types of generated code in the repo
and that contributors should not push any generated code themselves.

Signed-off-by: SAY-5 <say.apm35@gmail.com>