Update spec to use CDDL for data description.
In versions 1.1 and prior, the description for in-toto predicates was
informal pseudocode for specifying the shape of an object. With the
ratification of CDDL in RFC8610, it is possible to concisely describe an
object's regular structure in a standard language and furtheremore
account for more concise encoding with CBOR (from RFC8949).

To bring the base specification up to date with ITE-12, this change
constitutes a minor version increase to account for the new feature of
CBOR support. There is no need for a major version increase because the
JSON representation is semantically equivalent, given certain reasonable
interpretations of the meaning of the given pseudocode.