Create a simple verification attestation
While a VSA exists within the SLSA specification, the current tracks
have grown past the the initially proposed spec. We are proposing to
create a new attestation predicate to replace the VSA which is also
simpler to consume.

There is intentionally less information in this attestation. This was
partially done to create smaller predicates to enable faster
verification as well as removing informational fields. Even though
policy information is not present in the attestation itself, we call out
the ability to use a SLSA Provenance attestation to provide more
information about how the policy was evaluated. This provenance MAY be
used to qualify trust in the simple verification attestation if needed.

resolves: #277

Signed-off-by: arewm <arewm@users.noreply.github.com>