Add predicate type to the allowed media types for the envelope's payload type.
Predicate-specific media types are already recommended for individual
attestations in storage systems -- see
https://github.com/in-toto/attestation/blob/df7135a090fae2ed449bd95341652a9a2905b8e8/spec/v1/envelope.md#storage-convention.
This PR makes the equivalent change for the envelope's `payloadType`.

This change removes the inconsistency between the two uses of the media
types and makes it easier to consume attestations (e.g. when manually
opening up a bundle, https://github.com/in-toto/attestation/issues/405).
The risk is that a consumer uses solely the media type as evidence for a
policy decision, but this is warned against in the documentation.

A typo is also fixed in `bundle.md`.

Signed-off-by: Marco Deicas <mdeicas@google.com>